feat: Add Dockerfile and Makefile for dev environment
This commit adds a Dockerfile and Makefile to simplify the development environment setup for the project. The Dockerfile sets up a Node.js environment with the necessary dependencies and exposes the required ports for running the application. The Makefile provides convenient commands for building and running the Docker container.

The Makefile has been updated to include the ability to run the application with hot reloading using Vite. The Dockerfile has been updated to use `npm run dev` as the entrypoint command for starting the application.

The README file has been updated with instructions for using the Dockerfile and Makefile, and includes a separate section for each. Additionally, the README file has been updated to include a mention of the Makefile.

A fix has been added to the Makefile to ensure that the build command is run before the run command. An error related to `vite` not being able to find the dependencies for the application has been addressed with possible solutions in the README file.